hal: Fix alignement of the buffer sent to DSP
Currently, ALIGN macro can align properly only if the buffer size
is a power of 2.
To have same PCM samples for all channels, the buffer size requires to
be multiple of (number of channels * bytes per sample).
For writes to succeed, the buffer must be written at address which is
multiple of 32.
Alignment of 96 satsfies both of the above requirements.

Change-Id: I5ab33ef198b5b8f18ae9f09c4361e5cff84e700e
CRs-Fixed: 795936