hal: add use_small_bufs helper for offloading
Used to detect whether the buffers match a higher number of samples, in
this case they'll need to be shorter.
We don't have the code ready for any offload higher than 16bit on this
branch, so it will never need to use small buffers but we'll need to
compile the frameworks code part of all this.

Signed-off-by: Ícaro Hoff <icarohoff@gmail.com>