audio_extn: Implement handling for Cirrus Logic CS35L41 on Sony
Some new Sony devices have got a Cirrus Logic CS35L41 amp with
Cirrus Logic HALO DSP: this needs to be kickstarted in a special
way, calibrating it on the fly, loading the appropriate
firmwares for speaker protection and injecting the calibrations.

The current HAL "driver" will:
- Load the "Ambient Calibration" data from the TA partition,
  which is a fixed parameter that never changes during time
  and gets programmed to the HALO at each Calibration or
  Protection load;
- Check for /data/vendor/audio/sony_cirrus.cal file and:
-- If it exists, it'll check contents and - if ok - read the
   calibrations from this file and apply them at startup
-- If it doesn't exist, the calibration process will be
   started; it will take around 3 seconds from beginning to
   end. The user will hear a short sound.
   Once calibration is done and verified, the data will be
   saved to /data/vendor/audio/sony_cirrus.cal and will be
   loaded at every boot / every HAL restart.
- At every playback attempt (every attempt to use the DSP/AMP)
  do an error state verification on the DSP itself and:
-- If error detected, reset the DSP and crash the HAL, so that
   the entire initialization restarts. This is a very unlikely
   case that should basically never happen.
-- If no error detected, smooth sailing: just hands over to
   the standard Qualcomm HAL codepaths to start the playback

This software requires to read a boolean Android property
"vendor.audio.enable.cirrus.speaker" to decide whether it is
allowed to initialize or not; this is due to the fact that it
is injected into the Qualcomm HAL as a Speaker Protection
component: doing it this way allowed me to not touch the base
HAL code at all, apart one very small change that can even be
avoided (will come after this commit anyway).

Of course, this module supports two kinds of configurations:
- Mono speaker configuration, currently found in Sony Seine
- Stereo speaker configuration, currently found in Sony Edo.

Please remember that this HAL module requires support in the
kernel for the HALO DSP and the target audio amp and is also
dependant on mixer paths that are read from the firmware that
is loaded: if firmwares go different, this module must be
adjusted accordingly.

Change-Id: I62221f7a9e47b44c90d9653801bbba5248e4e4fb