cirrus_sony: Update path in adition to resetting in stop_processing
Just resetting the path at hand does not seem to recursively reset
included `<path>`s.

In the case of our dual mono feature `EAR_RDAC Switch` stays on when
headphones are plugged in after playing back audio on speakers meaning
that the sound is still coming from the earpiece.

With this change `EAR_RDAC Switch` is properly reset to `Off` when
plugging in headphones and audio does not play over the earpiece
anymore.

Fixes: 9c89b44a1fc747c8008d8be468d3ed214cf4c147
Change-Id: I9ba2b152baf96ca7c9520ff9518a0fae73f76004
Signed-off-by: Marijn Suijten <marijns95@gmail.com>