cirrus_sony: Do not sum error codes with positive string-length
cirrus_format_mixer_name returns the number of formatted characters
(result of snprintf). Returning this positive number is not always a
good idea, like when check_error_state returns 22 as the length of a
mixer control, which is caught by the `==0` check resulting in a full
reset.
Secondly, a postive length of 22 together with an actual, expected
failure of -EINVAL which conveniently is -22 results in 0; no error at
all!!

Change-Id: I5084e49e68b1bc78a08c551e940791e3a098308a
Signed-off-by: Marijn Suijten <marijns95@gmail.com>