hal: Parse compress metadata only if usecase is compress offload.
Parse Compress metadata only if usecase is compress offload, Set
isCompressMetadataAvail flag to true only if we parse the compress
codec related metadata so that we don't see the flag in any other
setParameter condition.
Ret value of audio_extn_parse_compress_metadata is always, so instead
of relying on ret value, pass the isCompressMetadataAvail flag to the
API, so that it can updated by audio_extn_parse_compress_metadata.

Change-Id: I8609991c753bb2a7e37c25c5a427af38c0f3fa9d