hal: check class config_ sample rate for HIFI Filter
using different class member for checking sample rate to
decide if HIFI Filter path need to be taken or not.

Previously sample rate member has not been updated yet in
the RouteStream when playback just start. Prior to 63M there
is no RouteStream call before the Open so it was fine.

Change-Id: I8e903a9f98a0d035a969e3c269bd8ffa3a7dc0f0