hal: Clear android devices if usb device is unavailable
Below scenario could occur:
Initially Tx-stream routed to built-in mic.
Later same stream is request over usb device, however, it
fails in RouteStream as usb-device may have plugged out.
At this point, mPalInDevice would point to usb-device
but mAndroidInDevices is not updated due to error, and
would still point to built-in mic. Due to this, if the
next request is over built-mic there won't be any device
switch due to same routing request, as mAndroidInDevices
points to same built-in mic.
And during in_read() it would still consider usb-device as
mPalInDevice is not updated, which throws error as usb-device
is unavailable.
The change is to clear mAndroidInDevices during error, such
that the next routing request is handled accordingly. 

Change-Id: I38bf488857e84d58e06c27de93ea92950c547ecb